requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
seaborn>=0.12.0
matplotlib>=3.7.0
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
tabulate>=0.9.0
openpyxl>=3.1.0
jupyter>=1.0.0
ipykernel>=6.25.0
pingouin>=0.5.3
krippendorff>=0.6.0
kaggle>=2.0.0
bleach>=6.0.0
